+++ src/Exception/RequestException.php
@@ -2,19 +2,19 @@
 
 namespace Fakturoid\Exception;
 
+use Fakturoid\Response;
 use Psr\Http\Client\RequestExceptionInterface;
 use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
 use Throwable;
 
 class RequestException extends Exception implements RequestExceptionInterface
 {
     public function __construct(
         private readonly RequestInterface $request,
-        private readonly ResponseInterface $response,
+        private readonly Response $response,
         ?Throwable $previous = null
     ) {
-        parent::__construct($response->getReasonPhrase(), $response->getStatusCode(), $previous);
+        parent::__construct($response->getOriginalResponse()->getReasonPhrase(), $response->getStatusCode(), $previous);
     }
 
     public function getRequest(): RequestInterface
@@ -22,7 +22,7 @@ class RequestException extends Exception implements RequestExceptionInterface
         return $this->request;
     }
 
-    public function getResponse(): ResponseInterface
+    public function getResponse(): Response
     {
         return $this->response;
     }